backend/requirements.txt
# Flask
Flask==2.2.2
Bcrypt
flask_bcrypt
itsdangerous==2.0.1
flask-cors
# Database
flask-mongoengine==0.9.5
flask_jwt_extended==3.24.1
flask_restful==0.3.8
Jinja2==2.11.1
MarkupSafe==1.1.1
